Revert "merge branch minh/week9 to main"
