+++ Models/DTOs/UpdateDentistAccountRequest.cs
@@ -9,7 +9,6 @@ namespace Models.DTOs
 {
     public class UpdateDentistAccountRequest
     {
-        [Key]
         [Required(ErrorMessage = "Email is required")]
         [EmailAddress(ErrorMessage = "Invalid email format")]
         public string? Email { get; set; }
