+++ Models/ProfessionDetail.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -9,7 +8,6 @@ namespace Models
 {
     public partial class ProfessionDetail
     {
-        [Key]
         public int ProfessionId { get; set; }
         public int DentistId { get; set; }
         public string DentistName { get; set; }
