+++ OData/DentalClinic_BE_OData/DentalClinic_BE_OData/Program.cs
@@ -1,7 +1,6 @@
 using Microsoft.AspNetCore.OData;
 using Microsoft.OData.ModelBuilder;
 using Models;
-using Models.DTOs;
 using Repositories.Implement;
 using Repositories.Interface;
 using Services.Implement;
@@ -11,11 +10,8 @@ var builder = WebApplication.CreateBuilder(args);
 // Add services to the container.
 var modelBuilder = new ODataConventionModelBuilder();
 modelBuilder.EntitySet<Dentist>("Dentists");
-modelBuilder.EntitySet<Schedule>("Schedules");
-modelBuilder.EntitySet<MedicalRecord>("MedicalRecord");
-modelBuilder.EntitySet<ProfessionDetail>("ProfessionDetails");
-modelBuilder.EntitySet<Profession>("Professions");
-modelBuilder.EntitySet<UpdateDentistAccountRequest>("UpdateDentistAccount");
+modelBuilder.EntitySet<Treatment>("Treatments");
+modelBuilder.EntitySet<Patient>("Patients");
 var edmModel = modelBuilder.GetEdmModel();
 
 builder.Services.AddControllers().AddOData(options =>
@@ -35,7 +31,7 @@ builder.Services.AddScoped<IRepositoryBase<Customer>, CustomerRepository>();
 builder.Services.AddScoped<IRepositoryBase<Patient>, PatientRepository>();
 builder.Services.AddScoped<IRepositoryBase<Schedule>, ScheduleRepository>();
 builder.Services.AddScoped<IRepositoryBase<Treatment>, TreatmentRepository>();
-builder.Services.AddScoped<IRepositoryBase<Transaction>, TransactionRepository>();
+builder.Services.AddScoped<IRepositoryBase<Models.Transaction>, TransactionRepository>();
 
 builder.Services.AddScoped<IAccountService, AccountService>();
 builder.Services.AddScoped<IPatientService, PatientService>();
@@ -47,7 +43,6 @@ builder.Services.AddScoped<IScheduleService, ScheduleService>();
 builder.Services.AddScoped<IScheduleService, ScheduleService>();
 builder.Services.AddScoped<IAppointmentDetailService, AppointmentDetailService>();
 builder.Services.AddScoped<IScheduleService, ScheduleService>();
-builder.Services.AddScoped<IMedicalRecordService, MedicalRecordService>();
 
 
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
